requirements.txt
pydantic>=2.5
pydantic-settings
PyYAML
requests
pdfplumber
sentence-transformers
scikit-learn